requirements.txt
Django==3.2.20
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6